SymbolCache: Move `clear_cache()` test to `test_symbol_caching.py`
