+++ tests/test_symbol_caching.py
@@ -41,3 +41,17 @@ def test_symbol_cache_aliasing():
     clear_cache()
     assert len(_SymbolCache) == 1  # We still have a reference to u_h
     assert np.allclose(u_h.data, 6.)  # u_h.data is alive
+
+
+def test_clear_cache(nx=1000, ny=1000):
+    clear_cache()
+    cache_size = len(_SymbolCache)
+
+    for i in range(10):
+        assert(len(_SymbolCache) == cache_size)
+
+        DenseData(name='u', shape=(nx, ny), dtype=np.float64, space_order=2)
+
+        assert(len(_SymbolCache) == cache_size + 1)
+
+        clear_cache()
